Refactor
Se modularizó el proyecto, separando cada modal en su archivo propio, también modularizando otros detalles del proyecto.
