+++ src/lib/translations.ts
@@ -1,0 +1,281 @@
+
+export const translations = {
+  es: {
+    hero: {
+      title: "Bienvenidos a Departamento KALEN",
+      subtitle: "Esperamos que disfruten su estadía"
+    },
+    categories: {
+      rules: "Reglas de convivencia",
+      checkin: "Check-in / Check-out",
+      wifi: "Wifi",
+      host: "Tu Anfitrión",
+      emergency: "Emergencias",
+      guide: "Guía Turística",
+      parking: "Estacionamiento Medido"
+    },
+    footer: "© 2026 Guía de Alojamiento",
+    comingSoon: "Disponible próximamente",
+    copy: "Copiar",
+    parking: {
+      title: "Estacionamiento Medido",
+      description: "La ciudad cuenta con sistema de estacionamiento medido.",
+      schedule: [
+        "Lunes a Viernes: 8 hs a 20 hs",
+        "Sábados: 8 hs a 13 hs"
+      ],
+      moreInfo: "Para mayor información ingresar a:",
+      link: "https://saemsmandes.com.ar/"
+    },
+    rules: {
+      title: "Reglas de la casa",
+      items: [
+        "No está permitido fumar/vapear.",
+        "No se permiten fiestas ni eventos.",
+        "Por favor, respetar el horario de silencio (23 a 7).",
+        "Cuidar utensilios y vajilla, en caso de faltantes y/o roturas, deberán ser abonados o repuestos.",
+        "La limpieza de la cabaña no incluye el lavado de vajilla.",
+        "El recambio de sábanas, toallas y toallones se realizará una vez a la semana, a solicitud del huésped, en estadías superiores a 7 noches.",
+        "La limpieza general se realizará una vez a la semana a solicitud del huésped -en estadías mayores a 7 noches. El mantenimiento diario queda a cargo del huésped, con los elementos disponibles en el departamento.",
+        "No desconectar el router del wifi.",
+        "No deslizar la traba de la cerradura electrónica. En caso de requerir asistencia para desbloquearla, el huésped deberá abonar un cargo adicional.",
+        "No modificar la posición de las llaves de paso del baño.",
+        "Se ruega cuidar los recursos, como el agua y la luz."
+      ]
+    },
+    checkin: {
+      title: "Check-in / Check-out",
+      checkinTitle: "Check-in",
+      checkinSub: "En adelante",
+      checkoutTitle: "Check-out",
+      checkoutSub: "Máximo",
+      instructionsTitle: "Instrucciones de salida",
+      instructions: [
+        "El egreso es autónomo. Dejar las **llaves de ingreso** al edificio sobre la mesa.",
+        "Dejar **toallas y toallones usados** en la bañera.",
+        "No olvidar **sacar la basura** y dejarla en el cesto ubicado en la vereda."
+      ]
+    },
+    wifi: {
+      title: "Conexión Wifi",
+      network: "Red (Network)",
+      password: "Contraseña (Password)",
+      //scan: "Escanea el código QR que se encuentra en la mesa del comedor para conectar automáticamente."
+    },
+    host: {
+      title: "Tus Anfitrionas",
+      superhost: "Superhosts",
+      email: "Email del Departamento",
+      availability: "Disponible de 08:00 a 22:00 para consultas generales.\n24hs para urgencias del departamento."
+    },
+    emergency: {
+      title: "Números de Emergencia",
+      addressTitle: "Dirección del departamento",
+      address: "Tte. Gral. Roca 938, piso 3, departamento 302\nEdificio Pequeña Comarca (entre calles Elordi y Belgrano)\nSan Martín de los Andes, Neuquén"
+    },
+    guide: {
+      title: "Guía Turística",
+      back: "Volver",
+      menu: {
+        benefits: "Beneficios",
+        gastronomy: "Gastronomía",
+        activities: "Actividades",
+        info: "Información útil"
+      },
+      benefits: {
+        title: "Beneficios",
+        items: []
+      },
+      gastronomy: {
+        title: "Gastronomía",
+        items: [
+          { name: "Próspero", type: "Restaurante", link: "https://maps.app.goo.gl/od3RoZE9LrfUbAPZ9", rating: "5.0", recommended: true },
+          { name: "Vieja Deli", type: "Restaurante • Café", link: "https://maps.app.goo.gl/qUJk5ppyGcVuYn54A", rating: "4.4", recommended: true },
+          { name: "Unser Traum", type: "Café", rating: "3.8", recommended: true, branches: [
+            { name: "Suc. San Martín", link: "https://maps.app.goo.gl/zEU6d2VRagr13GUh6" },
+            { name: "Suc. Roca", link: "https://maps.app.goo.gl/yinYKqgcAzVXYkgKA" }
+          ]},
+          { name: "La Costa del Pueblo", type: "Restaurante", link: "https://maps.app.goo.gl/TZ5jZ99n7a6Vf3Jn6", rating: "3.9" },
+          { name: "Pantera Bar Bistro", type: "Bar • Restaurante • Sushi", link: "https://maps.app.goo.gl/n1NfFaQxvWFLXuwL6", rating: "4.7" },
+          { name: "Dublin South Pub", type: "Bar • Café • Restaurante", link: "https://maps.app.goo.gl/emeAhms196HWYyu29", rating: "3.5" },
+          { name: "Ulises", type: "Restaurante", link: "https://maps.app.goo.gl/P1g7iWjiTw79Tm6m9", rating: "4.6" },
+          { name: "Piscis Parrilla", type: "Parrilla", rating: "4.1", recommended: true, branches: [
+            { name: "Suc. Mariano Moreno", link: "https://maps.app.goo.gl/6NCQBK8nLynEvKiL7" },
+            { name: "Suc. Belgrano", link: "https://maps.app.goo.gl/4dtidiWw2TEPbw3w7" }
+          ]},
+          { name: "La Vieja Aldea", type: "Chocolatería", link: "https://maps.app.goo.gl/BPJ7QcuvU8HHWbNv6", rating: "4.8", recommended: true },
+          { name: "Mamuschka", type: "Chocolatería • Heladería", link: "https://maps.app.goo.gl/xZom5UJynhuuySuT7", rating: "4.7" },
+          { name: "DELTURISTA", type: "Chocolatería • Heladería", link: "https://maps.app.goo.gl/NN158AZY5wz9ZJeEA", rating: "4.4" },
+          { name: "Pata Negra", type: "Chocolatería • Heladería", link: "https://maps.app.goo.gl/xkVBrrgz3ac5bvCGA", rating: "4.6" },
+          { name: "Riche", type: "Chocolatería • Heladería", link: "https://maps.app.goo.gl/M7crnZHtFCanuaTD8", rating: "4.3" },
+          { name: "Mamusia", type: "Chocolatería • Heladería", link: "https://maps.app.goo.gl/aACSwvDrruNzUWaw9", rating: "4.5" },
+          { name: "Grido", type: "Heladería", link: "https://maps.app.goo.gl/bcnxLZAnbNfV3hBx5", rating: "4.2" },
+          { name: "Heladería San Martín", type: "Heladería", link: "https://maps.app.goo.gl/gQFMm54HDXA8qzwe7", rating: "4.5" },
+          { name: "Pizza Cala", type: "Pizzería", link: "https://maps.app.goo.gl/NmgTeax3D138rvcw8", rating: "4.4" },
+          { name: "Genaro Pizza", type: "Pizzería", link: "https://maps.app.goo.gl/H2P7YkbPHVCmd7aK6", rating: "4.4" },
+          { name: "Gina Pizzeria", type: "Pizzería", link: "https://maps.app.goo.gl/wiyf5ZhwaWmgperx7", rating: "4.6" },
+          { name: "Forastera Café", type: "Café", link: "https://maps.app.goo.gl/E8vgZg28ncK5j2yJ9", rating: "4.8" },
+          { name: "Abolengo", type: "Café", link: "https://maps.app.goo.gl/DYuwzUmKrw3uD3Fa7", rating: "4.3" },
+          { name: "Fiora Café", type: "Café", link: "https://maps.app.goo.gl/TEMHySoe9hYgC26S8", rating: "4.4" },
+          { name: "Fassbier Cervecería", type: "Cervecería", link: "https://maps.app.goo.gl/HcWP9zVN1w9fdDj87", rating: "4.6" },
+          { name: "Cervecería Piedra Buena", type: "Cervecería", link: "https://maps.app.goo.gl/shPZobj7JEoAzQnS6", rating: "4.3" },
+          { name: "Jabalí", type: "Cervecería • Bar", link: "https://maps.app.goo.gl/Tg7Qd2rwpo6E5vvC8", rating: "4.1" },
+          { name: "Parador Slonjah", type: "Cervecería • Bar", link: "https://maps.app.goo.gl/UPHwZ4XhHmxnM2C2A", rating: "4.3" }
+        ]
+      },
+      activities: {
+        title: "Actividades",
+        items: [
+          { name: "Trekking al Mirador", difficulty: "Media", duration: "2 hs" },
+          { name: "Lago Lácar", difficulty: "Baja", duration: "Libre" }
+        ]
+      },
+      info: {
+        title: "Información útil",
+        items: [
+          { name: "Farmacia de Turno", desc: "Consultar cartelera en el centro" },
+          { name: "Supermercado", desc: "Abierto hasta las 21:00 hs" }
+        ]
+      }
+    }
+  },
+  en: {
+    hero: {
+      title: "Welcome to KALEN Apartment",
+      subtitle: "We hope you enjoy your stay"
+    },
+    categories: {
+      rules: "House Rules",
+      checkin: "Check-in / Check-out",
+      wifi: "Wifi",
+      host: "Your Host",
+      emergency: "Emergency",
+      guide: "Tourist Guide",
+      parking: "Measured Parking"
+    },
+    footer: "© 2026 Accommodation Guide",
+    comingSoon: "Coming soon",
+    copy: "Copy",
+    parking: {
+      title: "Measured Parking",
+      description: "The city has a measured parking system.",
+      schedule: [
+        "Monday to Friday: 8 AM to 8 PM",
+        "Saturdays: 8 AM to 1 PM"
+      ],
+      moreInfo: "For more information visit:",
+      link: "https://saemsmandes.com.ar/"
+    },
+    rules: {
+      title: "House Rules",
+      items: [
+        "Smoking/vaping is not allowed.",
+        "No parties or events allowed.",
+        "Please respect quiet hours (11 PM to 7 AM).",
+        "Take care of utensils and dishes; missing or broken items must be paid for or replaced.",
+        "Cabin cleaning does not include dishwashing.",
+        "Sheets, towels and bath towels will be changed once a week, upon guest request, for stays longer than 7 nights.",
+        "General cleaning is done once a week upon guest request -for stays longer than 7 nights. Daily maintenance is the guest's responsibility using available supplies in the apartment.",
+        "Do not disconnect the Wi-Fi router.",
+        "Do not slide the electronic lock latch. If assistance is required to unlock it, an additional charge will apply.",
+        "Do not modify the bathroom water valve positions.",
+        "Please be mindful of resources like water and electricity."
+      ]
+    },
+    checkin: {
+      title: "Check-in / Check-out",
+      checkinTitle: "Check-in",
+      checkinSub: "Onwards",
+      checkoutTitle: "Check-out",
+      checkoutSub: "Maximum",
+      instructionsTitle: "Departure Instructions",
+      instructions: [
+        "Self-checkout. Leave **building keys** on the table.",
+        "Leave **used towels and bath towels** in the bathtub.",
+        "Do not forget to **take out the trash** and leave it in the bin on the sidewalk."
+      ]
+    },
+    wifi: {
+      title: "Wifi Connection",
+      network: "Network",
+      password: "Password",
+      scan: "Scan the QR code on the dining table to connect automatically."
+    },
+    host: {
+      title: "Your Hosts",
+      superhost: "Superhosts",
+      email: "Apartment Email",
+      availability: "Available from 08:00 to 22:00 for general inquiries.\n24h for apartment emergencies."
+    },
+    emergency: {
+      title: "Emergency Numbers",
+      addressTitle: "Apartment Address",
+      address: "Tte. Gral. Roca 938, Floor 3, Apt 302\nEdificio Pequeña Comarca (between Elordi and Belgrano)\nSan Martín de los Andes, Neuquén"
+    },
+    guide: {
+      title: "Tourist Guide",
+      back: "Back",
+      menu: {
+        benefits: "Benefits",
+        gastronomy: "Gastronomy",
+        activities: "Activities",
+        info: "Useful Info"
+      },
+      benefits: {
+        title: "Benefits",
+        items: []
+      },
+      gastronomy: {
+        title: "Gastronomy",
+        items: [
+          { name: "Próspero", type: "Restaurant", link: "https://maps.app.goo.gl/od3RoZE9LrfUbAPZ9", rating: "5.0", recommended: true },
+          { name: "Vieja Deli", type: "Restaurant • Cafe", link: "https://maps.app.goo.gl/qUJk5ppyGcVuYn54A", rating: "4.4", recommended: true },
+          { name: "Unser Traum", type: "Cafe", rating: "3.8", recommended: true, branches: [
+            { name: "Br. San Martín", link: "https://maps.app.goo.gl/zEU6d2VRagr13GUh6" },
+            { name: "Br. Roca", link: "https://maps.app.goo.gl/yinYKqgcAzVXYkgKA" }
+          ]},
+          { name: "La Costa del Pueblo", type: "Restaurant", link: "https://maps.app.goo.gl/TZ5jZ99n7a6Vf3Jn6", rating: "3.9" },
+          { name: "Pantera Bar Bistro", type: "Bar • Restaurant • Sushi", link: "https://maps.app.goo.gl/n1NfFaQxvWFLXuwL6", rating: "4.7" },
+          { name: "Dublin South Pub", type: "Bar • Cafe • Restaurant", link: "https://maps.app.goo.gl/emeAhms196HWYyu29", rating: "3.5" },
+          { name: "Ulises", type: "Restaurant", link: "https://maps.app.goo.gl/P1g7iWjiTw79Tm6m9", rating: "4.6" },
+          { name: "Piscis Parrilla", type: "Grill", rating: "4.1", recommended: true, branches: [
+            { name: "Br. Mariano Moreno", link: "https://maps.app.goo.gl/6NCQBK8nLynEvKiL7" },
+            { name: "Br. Belgrano", link: "https://maps.app.goo.gl/4dtidiWw2TEPbw3w7" }
+          ]},
+          { name: "La Vieja Aldea", type: "Chocolate Shop", link: "https://maps.app.goo.gl/BPJ7QcuvU8HHWbNv6", rating: "4.8", recommended: true },
+          { name: "Mamuschka", type: "Chocolate Shop • Ice Cream Shop", link: "https://maps.app.goo.gl/xZom5UJynhuuySuT7", rating: "4.7" },
+          { name: "DELTURISTA", type: "Chocolate Shop • Ice Cream Shop", link: "https://maps.app.goo.gl/NN158AZY5wz9ZJeEA", rating: "4.4" },
+          { name: "Pata Negra", type: "Chocolate Shop • Ice Cream Shop", link: "https://maps.app.goo.gl/xkVBrrgz3ac5bvCGA", rating: "4.6" },
+          { name: "Riche", type: "Chocolate Shop • Ice Cream Shop", link: "https://maps.app.goo.gl/M7crnZHtFCanuaTD8", rating: "4.3" },
+          { name: "Mamusia", type: "Chocolate Shop • Ice Cream Shop", link: "https://maps.app.goo.gl/aACSwvDrruNzUWaw9", rating: "4.5" },
+          { name: "Grido", type: "Ice Cream Shop", link: "https://maps.app.goo.gl/bcnxLZAnbNfV3hBx5", rating: "4.2" },
+          { name: "Heladería San Martín", type: "Ice Cream Shop", link: "https://maps.app.goo.gl/gQFMm54HDXA8qzwe7", rating: "4.5" },
+          { name: "Pizza Cala", type: "Pizzeria", link: "https://maps.app.goo.gl/NmgTeax3D138rvcw8", rating: "4.4" },
+          { name: "Genaro Pizza", type: "Pizzeria", link: "https://maps.app.goo.gl/H2P7YkbPHVCmd7aK6", rating: "4.4" },
+          { name: "Gina Pizzeria", type: "Pizzeria", link: "https://maps.app.goo.gl/wiyf5ZhwaWmgperx7", rating: "4.6" },
+          { name: "Forastera Café", type: "Cafe", link: "https://maps.app.goo.gl/E8vgZg28ncK5j2yJ9", rating: "4.8" },
+          { name: "Abolengo", type: "Cafe", link: "https://maps.app.goo.gl/DYuwzUmKrw3uD3Fa7", rating: "4.3" },
+          { name: "Fiora Café", type: "Cafe", link: "https://maps.app.goo.gl/TEMHySoe9hYgC26S8", rating: "4.4" },
+          { name: "Fassbier Cervecería", type: "Brewery", link: "https://maps.app.goo.gl/HcWP9zVN1w9fdDj87", rating: "4.6" },
+          { name: "Cervecería Piedra Buena", type: "Brewery", link: "https://maps.app.goo.gl/shPZobj7JEoAzQnS6", rating: "4.3" },
+          { name: "Jabalí", type: "Brewery • Bar", link: "https://maps.app.goo.gl/Tg7Qd2rwpo6E5vvC8", rating: "4.1" },
+          { name: "Parador Slonjah", type: "Brewery • Bar", link: "https://maps.app.goo.gl/UPHwZ4XhHmxnM2C2A", rating: "4.3" }
+        ]
+      },
+      activities: {
+        title: "Activities",
+        items: [
+          { name: "Viewpoint Trekking", difficulty: "Medium", duration: "2 hrs" },
+          { name: "Lácar Lake", difficulty: "Easy", duration: "Free" }
+        ]
+      },
+      info: {
+        title: "Useful Info",
+        items: [
+          { name: "Duty Pharmacy", desc: "Check billboard in downtown" },
+          { name: "Supermarket", desc: "Open until 9:00 PM" }
+        ]
+      }
+    }
+  }
+}
